requirements.txt
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
PyYAML>=5.4.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0 